+++ src/components/Header.vue
@@ -1,0 +1,11 @@
+<script setup lang="ts">
+import { ref, reactive } from '@vue/reactivity'
+import { computed } from 'vue'
+import { useRouter } from 'vue-router'
+
+const router = useRouter()
+</script>
+
+<template></template>
+
+<style lang="css"></style>
