+++ src/main.ts
@@ -4,11 +4,11 @@
 // import BootstrapVue from 'bootstrap-vue';
 
 // createApp(App).use(BootstrapVue).mount('#app')
-import App from './App.vue'
-
 import { createApp } from 'vue'
+import App from './App.vue'
+import { createPinia } from 'pinia' // Import
 import BootstrapVue3 from 'bootstrap-vue-3'
-import router from '../src/router/index.js'
+import router from "../src/router/index"
 
 // Since every component imports their Bootstrap functionality,
 // the following line is not necessary:
@@ -19,5 +19,6 @@ import 'bootstrap-vue-3/dist/bootstrap-vue-3.css'
 
 const app = createApp(App)
 app.use(BootstrapVue3)
+app.use(createPinia())
 app.use(router)
 app.mount('#app')
